requirements.txt
torch>=2.0.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
tqdm>=4.65.0
scipy>=1.10.0
pyyaml>=6.0